+++ pkg/cypher/executor_use.go
@@ -206,6 +206,7 @@ func (e *StorageExecutor) cloneForStorage(store storage.Engine) *StorageExecutor
 	cloned.inlineEmbeddingChunkOverlap = e.inlineEmbeddingChunkOverlap
 	cloned.allowLocalAPOCImportFileAccess = e.allowLocalAPOCImportFileAccess
 	cloned.allowLocalAPOCExportFileAccess = e.allowLocalAPOCExportFileAccess
+	cloned.allowRemoteAPOCURLAccess = e.allowRemoteAPOCURLAccess
 	cloned.apocLocalFileAccessRoot = e.apocLocalFileAccessRoot
 	cloned.defaultEmbeddingDimensions = e.defaultEmbeddingDimensions
 	cloned.dbManager = e.dbManager